feat: OverrideDecalTextureWithMultiLayerImageCanvas
