+++ Editor/Inspector/Decal/AbstractDecalEditor.cs
@@ -9,47 +9,8 @@ using net.rs64.TexTransTool.Preview;
 
 namespace net.rs64.TexTransTool.Editor.Decal
 {
-    internal class AbstractDecalEditor : UnityEditor.Editor
+    internal static class DecalEditorUtil
     {
-        public override void OnInspectorGUI()
-        {
-            var thisSObject = serializedObject;
-            DrawerDecalEditor(thisSObject);
-
-            thisSObject.ApplyModifiedProperties();
-        }
-
-        public static void DrawerDecalEditor(SerializedObject thisSObject)
-        {
-            EditorGUILayout.LabelField("CommonDecal:label:RenderersSettings".Glc(), EditorStyles.boldLabel);
-            EditorGUI.indentLevel += 1;
-
-            var sTargetRenderers = thisSObject.FindProperty("TargetRenderers");
-            var sMultiRendererMode = thisSObject.FindProperty("MultiRendererMode");
-            TextureTransformerEditor.DrawerRenderer(sTargetRenderers, "CommonDecal:prop:TargetRenderer".Glc(), sMultiRendererMode.boolValue);
-            EditorGUILayout.PropertyField(sMultiRendererMode, "CommonDecal:prop:MultiRendererMode".Glc());
-
-
-            EditorGUI.indentLevel -= 1;
-            EditorGUILayout.LabelField("CommonDecal:label:TextureSettings".Glc(), EditorStyles.boldLabel);
-            EditorGUI.indentLevel += 1;
-
-
-            var sDecalTexture = thisSObject.FindProperty("DecalTexture");
-            EditorGUILayout.PropertyField(sDecalTexture, "CommonDecal:prop:DecalTexture".Glc());
-
-            var sColor = thisSObject.FindProperty("Color");
-            EditorGUILayout.PropertyField(sColor, "CommonDecal:prop:Color".Glc());
-
-            var sBlendType = thisSObject.FindProperty("BlendTypeKey");
-            EditorGUILayout.PropertyField(sBlendType, "CommonDecal:prop:BlendTypeKey".Glc());
-
-            var sTargetPropertyName = thisSObject.FindProperty("TargetPropertyName");
-            EditorGUILayout.PropertyField(sTargetPropertyName, "CommonDecal:prop:TargetPropertyName".Glc());
-            EditorGUI.indentLevel -= 1;
-        }
-
-
 
         static bool FoldoutAdvancedOption;
         public static void DrawerAdvancedOption(SerializedObject sObject)
